intersect ray and sphere + render

+++ include/physics/Vector3.h
@@ -8,25 +8,25 @@ using namespace std;
 class Vector3
 {
 public:
-    double x;
-    double y;
-    double z;
-    double length;
+    float x;
+    float y;
+    float z;
+    float length;
 
-    Vector3(double x=0, double y=0, double z=0);
+    Vector3(float x=0, float y=0, float z=0);
 
     Vector3 operator+(const Vector3& v) const;
     Vector3 operator-(const Vector3& v) const;
     Vector3 operator*(Vector3 v) const;
-    Vector3 operator*(double f) const;
-    Vector3 operator/(double f) const;
+    Vector3 operator*(float f) const;
+    Vector3 operator/(float f) const;
     void operator+=(const Vector3& v);
     void operator-=(const Vector3& v);
     void operator*=(Vector3 v) const;
-    void operator*=(double f);
-    void operator/=(double f);
+    void operator*=(float f);
+    void operator/=(float f);
     Vector3 normalize() const;
 
 private:
-    double _length() const;
+    float _length() const;
 };
