+++ src/physics/object/Sphere.cpp
@@ -1,4 +1,4 @@
 #include "physics/object/Sphere.h"
 
-Sphere::Sphere(const Point3& center, double radius)
+Sphere::Sphere(const Point3& center, float radius)
     : center(center), radius(radius) {}
