+++ src/rendering/Screen.cpp
@@ -1,5 +1,5 @@
 #include "rendering/Screen.h"
 
 
-Screen::Screen(Point3 origin, double width, double height)
+Screen::Screen(Point3 origin, float width, float height)
     : origin(origin), width(width), height(height) {}
